pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = "A Django project using layered architecture with form validation."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "myproject" },
    { include = "myapp" }
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14"
drf-yasg = "^1.21.5"
pydantic = "^1.10"
gunicorn = "^20.1"
pandas = "^1.5.3"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2"
pytest-django = "^4.5"
black = "^23.1"
isort = "^5.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
